+++ backend/api_business.py
@@ -1,9 +1,24 @@
 from fastapi import Request, Depends, status
 from fastapi.responses import JSONResponse
-from .models import UserProfile, SignatureResult, InstallationSummary, InstallationDetailsResult
-from .datasources import profile_info, sign_document, installation_list, installation_details
+from .models import (
+    UserProfile,
+    SignatureResult,
+    InstallationSummary,
+    InstallationDetailsResult,
+    Invoice,
+    InvoicePdf,
+)
+from .datasources import (
+    profile_info,
+    sign_document,
+    installation_list,
+    installation_details,
+    invoice_list,
+    invoice_pdf,
+)
 from .erp import ErpConnectionError
 from .auth import validated_user
+from .utils.responses import PdfStreamingResponse
 from consolemsg import error
 
 def setup_business(app):
@@ -33,3 +48,19 @@ def setup_business(app):
     def api_installation_details(contract_number: str, user: dict = Depends(validated_user)) -> InstallationDetailsResult:
         return installation_details(user['username'], contract_number)
 
+    @app.get('/api/invoices')
+    def api_invoice_list(user: dict = Depends(validated_user)) -> list[Invoice]:
+        return invoice_list(user['username'])
+
+    @app.get(
+        '/api/invoice/{invoice_number}/pdf',
+        response_class=PdfStreamingResponse,
+    )
+    def api_invoice_pdf(invoice_number: str, user: dict = Depends(validated_user)):
+        from yamlns import ns
+        result: InvoicePdf = invoice_pdf(user['username'], invoice_number)
+
+        return PdfStreamingResponse(
+            binary_data=result.content,
+            filename=result.filename,
+        )