+++ backend/datasources/erp.py
@@ -2,7 +2,15 @@ from contextlib import contextmanager
 from yamlns import ns
 from consolemsg import error, success
 from pydantic import ValidationError
-from ..models import TokenUser, UserProfile, SignatureResult, InstallationSummary, InstallationDetailsResult
+from ..models import (
+    TokenUser,
+    UserProfile,
+    SignatureResult,
+    InstallationSummary,
+    InstallationDetailsResult,
+    Invoice,
+    InvoicePdf,
+)
 from .. import erp
 from ..utils.gravatar import gravatar
 from ..utils.vat import nif2vat
@@ -15,9 +23,10 @@ from .exceptions import(
     NoSuchUser,
     NoDocumentVersions,
 )
+from .dummy import dummy_invoice_pdf
 
 @contextmanager
-def catchValidationErrors():
+def catch_validation_error():
     try: yield
     except ValidationError as exception:
         raise ErpValidationError(exception)
@@ -29,7 +38,7 @@ expected_erp_exceptions = [
     NoSuchUser,
     NoDocumentVersions,
 ]
-def processErpErrors(erp_response):
+def process_erp_errors(erp_response):
     if not 'error' in erp_response: return
     erp_errors = {
         excp.__name__: excp
@@ -42,35 +51,35 @@ def erp_user_info(login: str):
     e = erp.Erp()
     # TODO: Handle emails as login
     result = ns(e.identify(nif2vat(login)))
-    # TODO: processErpErrors(retrieved)
+    # TODO: process_erp_errors(retrieved)
     if 'error' in result:
         error(result.dump())
         return None
 
     result.avatar = gravatar(result.email)
 
-    with catchValidationErrors():
+    with catch_validation_error():
         return TokenUser(**result)
 
 def erp_profile_info(user_info: dict) -> UserProfile:
     e = erp.Erp()
     retrieved = e.profile(user_info['username'])
-    # TODO: processErpErrors(retrieved)
-    with catchValidationErrors():
+    # TODO: process_erp_errors(retrieved)
+    with catch_validation_error():
         return UserProfile(**retrieved)
 
 def erp_sign_document(username: str, document: str) -> SignatureResult:
     e = erp.Erp()
     retrieved = e.sign_document(username, document)
-    processErpErrors(retrieved)
-    with catchValidationErrors():
+    process_erp_errors(retrieved)
+    with catch_validation_error():
         return SignatureResult(**retrieved)
 
 def erp_installation_list(username: str) -> list[InstallationSummary]:
     e = erp.Erp()
     installations = e.list_installations(username)
-    processErpErrors(installations)
-    with catchValidationErrors():
+    process_erp_errors(installations)
+    with catch_validation_error():
         return [
             InstallationSummary(**installation)
             for installation in installations
@@ -79,9 +88,26 @@ def erp_installation_list(username: str) -> list[InstallationSummary]:
 def erp_installation_details(username: str, contract_number: str) -> InstallationDetailsResult:
     e = erp.Erp()
     retrieved = e.installation_details(username, contract_number)
-    processErpErrors(retrieved)
-    with catchValidationErrors():
+    process_erp_errors(retrieved)
+    with catch_validation_error():
         return InstallationDetailsResult(**retrieved)
+
+def erp_invoice_list(username: str) -> list[Invoice]:
+    e = erp.Erp()
+    invoices = e.list_invoices(username)
+    process_erp_errors(invoices)
+    with catch_validation_error():
+        return [
+            Invoice(**invoice)
+            for invoice in invoices
+        ]
+
+def erp_invoice_pdf(username: str, invoice_number: str) -> InvoicePdf:
+    e = erp.Erp()
+    pdffile=e.invoice_pdf(username, invoice_number)
+    process_erp_errors(pdffile)
+    with catch_validation_error():
+        return InvoicePdf(**pdffile)
 
 class ErpBackend():
     def user_info(self, login: str) -> TokenUser | None:
@@ -98,3 +124,10 @@ class ErpBackend():
 
     def installation_details(self, username: str, contract_number: str) -> InstallationDetailsResult:
         return erp_installation_details(username, contract_number)
+
+    def invoice_list(self, username: str) -> list[Invoice]:
+        return erp_invoice_list(username)
+
+    def invoice_pdf(self, username: str, invoice_number: str) -> InvoicePdf:
+        return erp_invoice_pdf(username, invoice_number)
+